Enhancement: Extract IsWritableTest
